feat: movendo botao wiki para componente
