+++ frontend/src/App.vue
@@ -3,19 +3,12 @@ import {
   onErrorCaptured,
   provide,
   ref,
-  watch,
 } from 'vue';
-import { storeToRefs } from 'pinia';
 import { Alert, EditModal, SideBar } from '@/components';
 import { useAlertStore } from '@/stores/alert.store';
 import { useAuthStore } from '@/stores/auth.store';
 import BarraDePendência from './components/BarraDeChamadaPendente.vue';
-import useRotaAtual from './composables/useRotaAtual';
-import { useWikiStore } from './stores/wiki.store';
-
-const { rotaAtual } = useRotaAtual();
-const wikiStore = useWikiStore();
-const { temWiki, wikiAtual } = storeToRefs(wikiStore);
+import BotaoWiki from './components/BotaoWiki.vue';
 
 const gblLimiteDeSeleçãoSimultânea = Number.parseInt(
   import.meta.env.VITE_LIMITE_SELECAO,
@@ -59,10 +52,6 @@ if (import.meta.env.VITE_COR_DA_FAIXA_DE_CONSTRUCAO || import.meta.env.DEV || ['
 
   window.document.documentElement.classList.add('dev-environment');
 }
-
-watch(rotaAtual, () => {
-  wikiStore.selecionarPaginaAtual(rotaAtual.value?.path);
-}, { immediate: true });
 </script>
 
 <template>
@@ -116,14 +105,7 @@ watch(rotaAtual, () => {
   <Alert />
   <div id="modais" />
 
-  <a
-    v-if="temWiki"
-    class="botao-wiki botao-wiki--direito-cima"
-    target="_blank"
-    :href="wikiAtual"
-  >
-    <svg><use xlink:href="#wiki_?" /></svg>
-  </a>
+  <BotaoWiki />
 </template>
 <style lang="less">
 @import url("@/_less/style.less");
